Ignore archived alumni roles in cleanup (hitobito/hitobito_jubla#223)
The AlumnusManager did not know anything about the archived feature, why
it lead to exceptions when it tried to cleanup obsolete alumnus roles.
Now, archived roles are igored during cleanup.

There could be other issues with archived roles, e.g. when creating an
alumnus role, which we do not yet investigate.